rename Sample1、Sample2、Sample3
