+++ SampleList/app/src/main/java/dev/seabat/android/compose/samplelist/navigation/Screen.kt
@@ -2,7 +2,7 @@ package dev.seabat.android.compose.samplelist.navigation
 
 sealed class Screen(val route: String) {
     object Home : Screen("home")
-    object Sample1 : Screen("LazyVerticalGrid")
-    object Sample2 : Screen("StaggeredGrid")
-    object Sample3 : Screen("FlowRow")
+    object LazyVerticalGrid : Screen("LazyVerticalGrid")
+    object StaggeredGrid : Screen("StaggeredGrid")
+    object FlowRow : Screen("FlowRow")
 }
