+++ SampleList/app/src/main/java/dev/seabat/android/compose/samplelist/screens/LazyVerticalGridScreen.kt
@@ -1,15 +1,19 @@
 package dev.seabat.android.compose.samplelist.screens
 
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -17,20 +21,21 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun Sample2Screen(
+fun LazyVerticalGridScreen(
     onNavigateBack: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     Scaffold(
         topBar = {
             TopAppBar(
-                title = { Text("Staggered Grid") },
+                title = { Text("LazyVerticalGrid") },
                 navigationIcon = {
                     IconButton(onClick = onNavigateBack) {
                         Icon(
@@ -43,38 +48,35 @@ fun Sample2Screen(
         },
         modifier = modifier
     ) { innerPadding ->
-        CustomStaggeredGrid(
-            items = (1..40).map { "Item $it" },
+        LazyVerticalGridExample(
             modifier = Modifier.padding(innerPadding)
         )
     }
 }
 
+@OptIn(ExperimentalFoundationApi::class)
 @Composable
-fun CustomStaggeredGrid(
-    items: List<String>,
+fun LazyVerticalGridExample(
     modifier: Modifier = Modifier
 ) {
-    Row(modifier = modifier.fillMaxWidth()) {
-        val column1 = items.filterIndexed { index, _ -> index % 2 == 0 }
-        val column2 = items.filterIndexed { index, _ -> index % 2 != 0 }
+    val items = (1..20).map { "Item $it" }
 
-        LazyColumn(modifier = Modifier.weight(1f)) {
-            items(column1) {
-                Card(modifier = Modifier
-                    .padding(8.dp)
-                    .fillMaxWidth()) {
-                    Text(text = it, modifier = Modifier.padding(16.dp))
-                }
-            }
-        }
-
-        LazyColumn(modifier = Modifier.weight(1f)) {
-            items(column2) {
-                Card(modifier = Modifier
-                    .padding(8.dp)
-                    .fillMaxWidth()) {
-                    Text(text = it, modifier = Modifier.padding(16.dp))
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(2),
+        modifier = modifier.fillMaxSize(),
+        contentPadding = PaddingValues(8.dp),
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+        horizontalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        items(items.size) {
+            Card(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .aspectRatio(1f),
+                elevation = CardDefaults.cardElevation(4.dp)
+            ) {
+                Box(contentAlignment = Alignment.Center) {
+                    Text(items[it])
                 }
             }
         }
@@ -83,6 +85,6 @@ fun CustomStaggeredGrid(
 
 @Preview
 @Composable
-fun Sample2ScreenPreview() {
-    Sample2Screen(onNavigateBack = {})
+fun LazyVerticalGridScreenPreview() {
+    LazyVerticalGridScreen(onNavigateBack = {})
 }
